requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Werkzeug==3.0.3
MarkupSafe==2.1.5
